restapi: Remove spaces in realm specification
The current basic authentication implementation adds spaces before
and after the equals sign that goes after the "realm"
specification in the "WWW-Authenticate" header. The RFC (see
http://www.ietf.org/rfc/rfc2617.txt) does not explicitly states
that spaces are not allowed here, but all the examples are without
spaces.

In addition the presence of spaces breaks some clients. Firefox,
for example, will not display the realm name in the popup used to
request the user name and password.

Python urllib2 doesn't work at all.

This patch removes those problematic spaces and fixes a minor
spelling issue.

Change-Id: Ida685483e5732fab2cacf96dae597a3199f3093d