core: Remove PowerMockito from AbstractQueryTest
Removed the usage of PowerMockito from AbstractQueryTest in order to
improve tests performance. PowerMokcito was (mainly) used to mock
DbFacade.getInstance(), which is no longer needed since
getQuery.getDbFacade() can be spied.

Note that this pactch does NOT completely remove PowerMockito from all
the queries' tests. For queries that require mocking static functions
in helper classes (e.g., ImagesHandler, VmHandler), PowerMockito is
still required, and the further refactoring requried to get rid of it is
left for a future patch.
In some classes PowerMockito calls were left untouched, and in some
they had to be added explicitly since the base class no longer provides
this functionallity.

Change-Id: I3610e8ad9270d14a66475c67dd78cf4f8d1e64f0