webadmin: Gluster Add Bricks dialog changes
This dialog will allow the user to select bricks from the list of available bricks.
As opposed to initial design (single table with mutli-selection) this dialog will have 2 tables, one for showing
available bricks and another for showing the selected bricks. Add/Remove/Add All/Remove All
actions are provided. Move Up/Move Down actions are provided to change the order of
the selected bricks, as the order is impotant for Replicate/Stripe volumes.

Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>
Change-Id: I0fe600b6ae0dbc16d349af602d72febb6de22a4b