core: rename upgrade functions in db upgrade scripts
Removing the version suffix from function names in upgrade scripts.
The reasons are
1) Each time that you have to rename your upgrade script after sync with master you have to rename the function as well
2) As a result, may function are pushed with a wrong suffix (for example upgrade 03_01_0950 had a function name suffix of 08200)

Since the reason for that numbering was in the first place for avoiding dropping real installed functions, now you will have just to
put in a temporary function created in an upgrade script just the prefix "__temp_"

Hope that this will make handling upgrade scripts more accurate and less painful.

Change-Id: I73db2cf50776488b603d62afcee4033573c6092a
Signed-off-by: Eli Mesika <emesika@redhat.com>