webadmin: adding externalize check to the frontend checkstyle
no bz

1. Separating build-tools into two packages-
  * checksyles- for checkstyle xml files.
  * ovirt-checkstyle-extension- for adding to new (not built-in)
    checkstyle checks.
2. Add to the ckechstyle.xml new check- NlsCheck.
This check fails if there is a non externalized string in a java file and
there is no //$NON-NLS-N$ comment to mark that it is ok that the string
is not externalized.

To understand more about the need of this check
see http://gerrit.ovirt.org/#change,3612

Important- The NlsCheck wouldn't run by default. If you want
this check to run on a specific project you have to add
<propertyExpansion>runNlsCheck=false</propertyExpansion>
to the configuration of the checkstyle in the project's pom.

Signed-off-by: Alona Kaplan <alkaplan@redhat.com>
Change-Id: Ie23035701398158b70b61c1c85a7f6eb40770784