core: Introduce allow snapshot attribute for disk
Adding field of allow snapshot to disk entity.
For now this field will be used for shared disk and direct lun.
In the future it should indicate which disk will be snapshot in the VM
or not.
Disk which will have false allow snapshot attribute will be static in
the VM, which means there will be no representation in time of the disk.
When VM will be previewed and commited or rollbacked after that, the
disk will still be part of the VM.

Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Change-Id: I99caa62cd6f966fe6867fb2f42bcf5019fa65ee5