webadmin: Support for Stop Volume
Stop volume option will be disabled if the volume is already down.

When stop volume option is clicked, a window would popup asking for confirmation of stop volume.
Suppose multiple volumes are selected some of which are up and some are down,
both start and stop option will be disabled.

Change-Id: I0e799d24d79915ac0c87bc9c7a0c4e9275a65743
Signed-off-by: Raksha Rao <rrao@redhat.com>