core: VM status is updated although the last operation on it has not finished yet
Problem was that Resume command adds the vm to the _asyncRunningVms collection in ResourceMnager.
Then in the canExecute method of CreateVdsCommand the vm is checked for being in that collection and if it is
found , than the status of the VM is set to UP before the command even starts to execute.

The solution is to take and return the curent status as reflected in the vm_dynamic table.

Change-Id: I1e97372d23dd4edcba5a0e9147beabef81bc70e2
Signed-off-by: Eli Mesika <emesika@redhat.com>