core: Set parameters task success only for command
Set the task success only for the commands that actually failed, by
default tasks are considered to have ended all successfully.

The success is an aggregation of the status of the command itself
(would change if it also sent a task) and the status of all son
commands, where if one failed then as a whole the parent is considered
as failed.

This will allow to selectively fail the parent command and all son
commands that failed, while the son commands that have finished can
clean up after themselves, instead of leaving garbage.

This change will not harm current behavior, however, since the
VmCommand will set the son commands to failed if the parent command
has failed either way.

Change-Id: I4366b6606ae83064b313aed17b4f65384e22342b
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>