userportal: clone VM create VM with wrong permissions(#809778)
https://bugzilla.redhat.com/809778

Extended->Virtual Machines->select one->Snapshots->select one->clone

It created a new VM but invisible for the user which created it.

The problem was in VmSnapshotListModel.OnCloneVM() which
did not mark the creator to be the explicit owner
(AddVmFromSnapshotParameters.setMakeCreatorExplicitOwner())

Solved by creating a snapshot model specially for user portal
(UserPortalVmSnapshotListModel) which sets this flag to true.

Change-Id: I25aa6e78a4c6d7d516a97a910e5d57e4a0447341
Signed-off-by: Vojtech Szocs <vszocs@redhat.com>