core: New queries to get ISOs by StoragePoolIds
Introduced two new queries to get ISO and Floppy RepoFileMetaDatas
according to the Storage POOL they are contained in.

This patch extracts the common functionallity from
AbstractGetAllImagesListQuery and allows overriding the
getStorageDoaminId() method.
For the new queries, this is done under the assumption there is but one
ISO domain per storage pool.

Change-Id: I86add10f226074792966d6d3efcc67499e459565
Signed-off-by: Allon Mureinik <amureini@redhat.com>