core: db - adding a flag for upgrade to clean tasks metadata
Adding a -c flag to the upgrade script to clean data in async_tasks and
Business_entity_snapshot tables.
This fixes the last commit on that issue that cleans this data in each upgrade.
This is pre step for an installer fix that will check before upgrade if there are
running taks and if so, asks the user to wait for its completion.
The installer will supply a checkbox to mimic the call to the upgrade script with the -c
flag and force cleaning that data.

Change-Id: I946fc1c4f1bbcf4f3276e330721cc0b575fc961b
Signed-off-by: Eli Mesika <emesika@redhat.com>