core: after running vm as stateless we can no longer run the vm
Was unable to reproduce the bug.
Reported scenario worked in both run and run-once no matter if VM was stateless and
stateless set to false in run-once or vice-versa.

Fixied a bug in stateless VM unmanaged devices cleanup that were cleaned up to now only when VM was
suspended and then stop rather than on stop from any state as it should be.

Change-Id: Ia139e98596715bb3b09a589da6dfae98a38b7249
Signed-off-by: Eli Mesika <emesika@redhat.com>