core: Introducing MockConfigRule
Introduing MockConfigRule, as an easy way of mocking Config in order to
avoid the dependency on PowerMock that many tests have.

LdapSearchQueryTest is refactored to use this new @Rule as a usage
example.
If this strategy will get the community's approval (i.e., the patch
will be approved), other tests that can benefit from this @Rule will be
refactored in separate patches.

Change-Id: I9d5d66b135c07b6d6e61fb70bc477d2171cfcb22
Signed-off-by: Allon Mureinik <amureini@redhat.com>