core: split configuration common stored procedure
This commit adds a new "split configuration" function, called
fn_db_split_configuration. The purpose of this function is to easy
upgrading in case a configuration value that was once general, should
now be version-specific.

The input of the function is the option_name and the required new value.
The result is having the current value for each possible version,
besides the last version, which will get the input value.

Change-Id: Ic5f828b140d9ced9675e5b9091ee27f5a6d9ffe7
Acked-by: FirstName LastName <email>
Signed-off-by: Oved Ourfali <oourfali@redhat.com>