core: Moved LdapSearchQueryTest out of itest
LdapSearchQueryTest is not an integration test, it uses mocking to
simulate LDAP resources and tests the SearchQuery mechanisms. This patch
moves it out of itests and in to the org.ovirt.engine.core.bll package,
so it's in the same package as the class it tests.

This refactor also allowed to reduce the visibility of some helper
methods introduced in http://gerrit.ovirt.org/#change,4183.

Change-Id: I237929be0f837ef69b51be2205b08043d82e8fab
Signed-off-by: Allon Mureinik <amureini@redhat.com>