core: Add console reconnect
This patch adds the functionality to allow an user to connect to
the console of a virtual machine even if another is already
connected.

In order to add this functionality the following columns have been
added to the vm_dynamic table:

console_user_id
  The UUID of the last user that successfully connected to the
  console of the virtual machine.

This functionality will be enabled for server machines by default.
For desktops it is disabled by default and a new flag has been
added to enable/disable. In order to store this flag the following
column has been added to the vm_static table:

allow_console_reconnect
  Flag indicating if an user can connect to the console even if
  another user is already connected.

Change-Id: I9b4725e51c9fb7070920c77e4f8ae54c5af4377b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>