core: Support Sysprep for 3.1 and up (#811587)
https://bugzilla.redhat.com/811587

The patch uses the VM Payload to provide the Sysprep file content to the
VDSM when running a VM for cluster level 3.1 and up.

Floppy device isn't created implicitly in 3.1 therefore
engine should ask the agent to create one and treats it as unmanaged
device so it won't be persistent to the engine-db, and treated as
stateless.

Change-Id: I9c1f7f3edcc3aea7e38b3c6a7efdfebdd24269f1
Signed-off-by: Moti Asayag <masayag@redhat.com>