core : overloading engine with tasks will cause bin to get stuck with exception (#813817)
https://bugzilla.redhat.com/813817

The problem was caused because of the VdsBroker statless bean was opened for too much time.
The bean is resposnsible for encapsulation of connection to vdsm.
There are no any reasons to use a bean for the following case, that's why as a solution for bug I removed a
bean

Change-Id: Ia641e5348b9bda51dffadab2aae93719f58e975b
Signed-off-by: Michael Kublin <mkublin@redhat.com>