core: fix split function and its use for spice security channels
Previous commit added the split function, but it didn't take into
account what if the entries are not there at all (new install).
It assumed they are there, as they would appear in the add section.
Better implementation is not to assume that, creating the old entries
with old value in install phase, or taking the original value from the
DB in case of upgrade.

Change-Id: I32b9b89c0c28180738b8532561e110fcd3e61d1b
Signed-off-by: Oved Ourfali <oourfali@redhat.com>