core: Add validation to MAC format (#756437)
https://bugzilla.redhat.com/756437

The MAC address needs to be in a specific format, so adding a
validation so that when network is created/updated then the MAC
address format is correct.

The validation of MAC size is no longer required, since this new
validation will cover also shorter/longer MAC cases.

A dependency on slf4j-nop was added to the pom.xml file for scope test
to allow testing the validations, since the default validation provider
uses slf4j so we must provide some implementation.

Change-Id: I8d25cc5c48fb2ab48c9137cd167a43459cf03931
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>