core: Introducing RandomUtilsSeedingRule
The first revision of RandomUtilsSeedingRule, intended to allow logging
and overriding RandomUtils' seed for unit tests.

This patch includes the new @Rule itself and a usage example.

If this approched is accepted by the community and the patch is
merged, I'll submit further patches that utilize this rules in tests
that use RandomUtils.

Change-Id: I9b61902aca7986752e61530dc4afb1b9a9293c2d
Signed-off-by: Allon Mureinik <amureini@redhat.com>