core : Start to remove internal drive mapping
THe follwoing patch is a beginning of the remove of internal drive mapping property,
as first stage a clean up was done at AbstractDiskVmCommand.
The clean up at AbstractDiskVmCommand will also fix a bug, which will not allow to attach a shared disk to vm.
Also a clean up contains a removing of un needed pass of parameters at different images flows

Change-Id: I020b25f04820b6cbef60764ae8169050bcbdc5bc
Signed-off-by: Michael Kublin <mkublin@redhat.com>