packaging: Move ROOT.war to the EAR
This patch moves the root web application to the EAR file. This
has several advantages:

1. There is only one deployment unit to add to the application
server, so the setup and packaging is simplified.

2. The root web application works without needing to use the
engine-setup tool. This can be very useful for development
environments.

3. The root web application is a regular maven subproject so it
can be built taking advantage of all the maven capabilities. In
particular the content can be dynamically generated using
servlets. In fact this is already used to generate the following
files:

web-conf.js
  Instead of copying this file from the /etc/ovirt-engine
  directory the application contains a servlet that reads that
file
  and sends it content. It also can guess the content of the file
  even if it doesn't exist (useful for development environments).

ca.pem
  Instead of copying this file from the /etc/pki/ovirt-engine
  directory the application uses a servlet to read it from its
  original location and send the content with the right content
  type.

engine.ssh.key.txt
  Instead of copying the file from /etc/pki/ovirt-engine/keys
  directory the application uses a servlet to read it from its
  original location. This could be improved to generate the
  content directly from the keystore.

This means three files that don't need to be copied or symlinked
during the setup of the engine and can be placed only in the
directories they are expected to be according to the file system
standards (configuration files only in /etc, so on).

This patch also simplifies the setup tools so that they don't need
to modify the web.xml file.

Change-Id: I1d3aa9af7b78546b5449031c8db23ff766b68496
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>