api: Changing address to be optional
In order to conform to previous usages of
AddStorageDomain via REST-API, and due to the fact
that NFS is posix FS compliant, the user will be allowed
to use the "Address" and "Path" fields (a mountSpec of
address:path will be passed to VDSM).
However, as mountSpec is not required to have ":"
it will be possible to use path field only.

Change-Id: I52f9c73694952bc7839124e23d74d1b029dc6a6c
Acked-By:
Signed-off-by: Yair Zaslavsk <yzaslavs@redhat.com>