webadmin: Edit Pool shows different template (#820231)
https://bugzilla.redhat.com/820231

When creating a pool from a template, than edit pool
shows a different template than the one from which it has been
created.

The problem was, that the edit pool dialog used the same behavior
class for both edit and new pool (the NewPoolModelBehavior),
and this one selects always the first pool.

The fix was to create a separate class ExistingPoolModelBehavior
which initializes the template correctly.

Change-Id: Iea1270499ebd3b1c6f7ee93d62cbc7a8167fb7c1
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>