webadmin: New Template dialog: remove OK when dialog disabled (#589518)
https://bugzilla.redhat.com/589518

virtual machies main tab -> select a machine -> click "Make Template"

If there is an error which prevents the Template to be created, the
whole dialog gets disabled, however, the OK button is still present.

The logic of disabling of the button is in
NewTemplateVmModelBehavior.DisableNewTemplateModel() which missed
the removing of the OK button.

Added removing of the OK button to this method.

Change-Id: Ibadf4777ef6c4903e81920dbf62c67d336a5e025
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>