webadmin: Support for mode specific Tabs and Sub Tabs
This patch is part of Mode Specific UI implementation in Ovirt.

Currently supported modes are
 1. VirtOnly(1) --    binary equivalent  0000 0001
 2. GlusterOnly(2) -- binary equivalent  0000 0010
 3. VirtGluster(3) -- binary equivalent  0000 0011
 4. AllModes(255)  -- binary equivalent  1111 1111

The value for the UiMode is read at the time of loading and stored in ApplicationModeHelper.
UiMode class is moved from webadmin to uicommonweb

A new field(availableInModes) is introduced in Model class to specify in what modes the model
should be available. The value for this field is set to AllModes(255) by default. So that any
model become available in all modes by default. Appropriate values can be set in the respective
subclasses.

getIsAvailable() method of Model class is intercepted to check for the mode as well.

GlusterModeEnum and its references are removed.

Change-Id: I55da9510a1480634ce30cf77b9bd9ecb263186bc
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>