core: Support floating lun in the DAO
Before this patch DiskDAO would crash when attempting to read a floating
lun due to missing treatment for vm_entity_type.
Such a treatment existed in DiskImageDao, however.

This patch extarcts the AbstractBaseDiskRowMapper class, and changes
AbstractDiskRowMapper to handle Disks instead of BaseDisks.
Having done that, AbstractDiskRowMapper can now handle the entityType
field, so that both DiskImageRowMapper and LunDiskRowMapper that extend
it will get that functionality.

In addtition, this patch also adds a floating lun to the fixtures.xml to
test this functionality, and updates the other disks' tests accordingly.

Change-Id: Iec20c0092375cb1a7ee4b7b40f57e75d2c5cd13c
Signed-off-by: Allon Mureinik <amureini@redhat.com>