restapi: GET vm nic/disk does not work
In a recent change, an interface method with a generic return value
was overriden in an inheriting interface, and the generic return
value was replaced with the concrete return value. Rest-easy seems to
have difficulties with this.

This patch overrides the method in the implementing class and calls
super(), which seems to resolve the ambiguity for rest-easy.

Change-Id: Id1df3c20e593cd78b022d5b26582a1da8f4043f9
Signed-off-by: Ori Liel <oliel@redhat.com>