core: make the getTimeZone query OS bounded
this query will return different timezone for windows OS (the default
and old behaviour) otherwise the java timezone lists
as maintained in java.util.TimeZone see http://doc.java.sun.com/DocWeb/api/java.util.TimeZone.

purpose of this change is to support setting timezone for all VMs.
note: the timezone is used to set the hardware clock of the virtual bios and
not the OS timezone, which is done by sysprep for windows and future
sysprep-like for linux.

Change-Id: Ie4389f058d0a451ba4f7b9e50daab8fab87df450
Signed-off-by: Roy Golan <rgolan@redhat.com>