core: is_inheritable property to MLA entities
Added the is_inheritable property to the Action Group and Role MLA
business entities.

This patch handles the addition of the property itself in the java and
database sides.
It does not contain the changes in the logic that update or use this
property, which will be added in a later patch.

Change-Id: I8732306c515bade05db5cb95b9d571003fdec177
Signed-off-by: Allon Mureinik <amureini@redhat.com>