core:Support remove of shareable disk.
The following patch has the following changes:
1.Adding support to remove disk from multiple vms.
The remove command was counting on the vm id which was in the disk.
After this change the vmId will be fetched from the DB.

2.Update vm device store procedures to use vmId as a parameter when remove
or getting a vm device.
This is neccessery for the Activate/Deactivate process to work for
shareable disk.

3.Update the lock mechanism in the remove disk command to log the VMs
which are using in the shared lock, and also change the aquire lock to be at
the end of the CDA instead before the validation.

Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Change-Id: Idde509a7bdee94e6fd481fce510b26f89d6b0429