core: Disable application server welcome content
The application server welcome content is automatically disabled
by the engine-setup tool, but not in development deployments. Now
that the root web application has been moved to the .ear directory
we can end up having two root web applications. As a consequence
the application server will fail to deploy the .ear. This patch
changes the default configuration to disable its welcome content.

Change-Id: I01f3cc9a5542afc2f4c1649c968f575bb096ff66
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>