userportal, webadmin: when login automatically, disable login screen
When the user is logged in, and opens for example a new tab,
the application loggs him/her in automatically.

The problem is, that the login screen is not disabled during this
process. So it can happen, that during the process of loggin in
automatically, the user fills in some data into the login form,
which has no meening.

The solution was to disable the login screen and not enable it,
when it is in process of auto login.

Change-Id: I3b2a15eb8d17a00f07590996f96f617014e0ecd6
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>