core: Use same ids for artifacts and plugins
Currently we use different identifiers for the repositories used
for normal artifacts and the repositories used for plugins. This
may be triggering a bug in maven.

Apparently maven 3 doesn't consider artifacts downloaded from
different repositores to be the same, even if they use the same
gid:aid. Maven creates a _maven.repositories where it stores the
origin of the artifact like this:

  checkstyle-5.0.jar>id_of_the_repository=
  checkstyle-5.0.pom>id_of_the_repository=

For example, when the checkstyle artifact is downloaded from the
repository with id "maven1.central" the file will contain the
following:

  checkstyle-5.0.jar>maven1.repo=
  checkstyle-5.0.pom>maven1.repo=

Later when the maven-checkstyle-plugin is downloaded it comes from
the repository with id "central" and apparently it looks for the
checkstyle artifact from repository "central" as well, and it
fails.

This patch changes the root POM to use the same "centra"l"
identifier for both artifacts and plugins.

Change-Id: I5b0236f1dd0a1a18e226a22ed3563c462af90444
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>