core: Cleanup SetupNetworksHelper part 2.5
Change lazy initialization to be on the Maps, since the Lists are not
really used afterwards. This also required to change all calls to the
Maps to go through getter methods.

In turn, it is no longer necessary to pass the map around, as it is now
a state within the class, so removed it from all the calls (tests were
fixed accordingly).

Change-Id: I0453bdae390f1c5d543f5ff6423eec130ed8f60d
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>