core: Allow admin permissions in user views
The user permissions views are intended to show all the objects a given
user has permissions on, regardless of if it's an admin role or a user
role (e.g., a user who is an admin on his own VM should be able to see
itin the user portal).

This patch fixes this issue.

Change-Id: I27e60a3e855cb78a2cb773fdc53c3d1f35d8f058
Signed-off-by: Allon Mureinik <amureini@redhat.com>