userportal: SpiceUsbAutoShare configuration is ignored (#767081)
https://bugzilla.redhat.com/767081

While "SpiceUsbAutoShare" is set to true in vdc_options, the
Auto-Share USB checkbox in the user portal is not checked by default.

The problem was, that the SpiceUsbAutoShare was completly ignored.
Solved by adding the reading of this value to the
Configurator.updateSpiceUsbAutoShare() method which is called from the
constructor of SpiceConsoleModel via of the Configurator.Configure().

In most cases this value will be read from the cache, but for the
first time it will really reach out to the DB and it can take a
while. That is the reason, why the ConsolePopupPresenterWidget
listens to UsbAutoShareChangedEvent and re-renders the popup
if such an event arrives.

Change-Id: Iaf72cb10aeaf0c7bafca17ed91ac0fd330501d80
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>