core:Refactor locking mechanism
Refactor lock mechanism by swithing the key with the value.
Instead using entity name as key and guid as value, now the key will be
the GUI and the value will be the entity name.

This fix should be since now with shareable disk, there can be multiple
VMs which can be locked in one operation.

Change-Id: Icbe9e0eb5a0ec2e8af899b3e554b1cb6fccc6835
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>