engine,webadmin: oVirt Modes - Event Notifier changes
Event Notification List in Add Event Notification dialog
should be shown according to the mode. For example Gluster
related events should not be shown if the application is running
in VirtOnly mode.

Following changes are done to make this work:

1. EventNotificationEntity is modied to suuport mode specifications.

2. UiMode is renamed to ApplicationMode and moved to common module
as it is required in EventNotificationEntity

3. A new utility method is added to ApplicationModeHelper to filter
the event notification list according to the mode.

Change-Id: I5ec5840b1e2a78fb5f7b49280eb6b13ace74b15a
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>