core: Fix handling same network on different nics
The old check would never catch this scenario, since the
networksOverBond list was locally initiated to an empty list.

The check should be simple - network can be defined only once on some
interface.

Clarification:
If it is a bond interface then it can't be defined on the
bond slaves.
If it is a VLAN then it can't be defined on the VLAN's NIC.

Change-Id: I4cb7f7c0c7cc7c2240b72e411d81c2ddbc561766
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>