core: fix jobProperties for delete vm and template from export (#818111)
https://bugzilla.redhat.com/show_bug.cgi?id=818111

when removing vm/template from export domain,
the job properties tried to use entities from db,
which are not always available, and not correct to use.

this fix implement override for getVm() and getVmTemplate() to get from export domain,
and use it.
most of the code was just moved from canDoAction

also set the domain name which was not set,
and fixed in ExecutionMessages.properties to the right var name

Change-Id: I936501874e9a29b18bdddf1b65e7150e88cfe475
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>