restapi: error getting VM NIC with non-exists network (#782437)
https://bugzilla.redhat.com/782437

There are scenarios in which the network of a VM Nic does no longer
exists (e.g. exporting a VM template with nic, the network is removed
from the data-center and the template is imported, or using the same
export domain across various data-centers with different DC networks).

The patch removes the failure on retrieving the network associated with
a nic, if such network doesn't exist.

Change-Id: I66e94f92ead04a565bacb417b630db365ae4b682
Signed-off-by: Moti Asayag <masayag@redhat.com>