engine: Refactor LdapProviderType and LdapVendorNameEnum
This patch is still a work in progress. Its not complete and there are
holes in it. Like when merging LdapProviderType and LdapVendorNameEnum
I did not see a Vendor Name with AD. Need to know what goes there. Also
need to know what to do with 'general'.

I am emailing it out early to get feedback. There will be more versions
of it out based on feedback.

Roy and Yair, Is this what you were thinking?

Change-Id: I9d9e8ba5f7ada715c9016450d231177b2c69e2fb
Signed-off-by: Sharad Mishra <snmishra@linux.vnet.ibm.com>