core: UnPowerMock GetVmByVmIdQueryTest
Removed PowerMock dependecy from GetVmByVmIdQueryTest.

Before this patch, the test was not a good unit test - it did not test
just the query, but also VmHandler's functionallity, which should have
been tested separately.
This patch turns the test into a strict unit teset, which only test's
the query's functionallity, and this PowerMock is no longer needed.

Change-Id: Ieaf73c28f2321e564d9c9d336dbf33a89e7a2929
Signed-off-by: Allon Mureinik <amureini@redhat.com>