core: adding alias to vm_device
This patch adds the alias of the device to the vm_device table.
We use this field to sort devices in cases devices from the same type
appear more than once, and the order is important. We set the order
according to the alias. In this patch it is used mostly for video
devices.

Change-Id: I9d543dbb96b002defbe2b85c88a4f24fab8e5f04
Signed-off-by: Oved Ourfali <oourfali@redhat.com>