core: Java 7: Fix use of namespaces in StAX writer
The way we currently use namespaces in the StAX writer for OVF
files doesn't work with Java 7. When trying to write attributes
without a namespace the following exception is generated:

java.lang.RuntimeException: Failed to write attribute
  at org.ovirt.engine.core.compat.backendcompat.XmlTextWriter.WriteAttributeString(XmlTextWriter.java:44)
  at org.ovirt.engine.core.utils.ovf.OvfWriter.WriteHeader(OvfWriter.java:49)
  at org.ovirt.engine.core.utils.ovf.OvfWriter.<init>(OvfWriter.java:40)
  at org.ovirt.engine.core.utils.ovf.OvfVmWriter.<init>(OvfVmWriter.java:26)
  at org.ovirt.engine.core.utils.ovf.OvfManager.ExportVm(OvfManager.java:24)
Caused by: javax.xml.stream.XMLStreamException: NamespaceURI cannot be null
  at com.sun.xml.internal.stream.writers.XMLStreamWriterImpl.writeAttribute(XMLStreamWriterImpl.java:611)
  at org.ovirt.engine.core.compat.backendcompat.XmlTextWriter.WriteAttributeString(XmlTextWriter.java:42)

This patch changes the OVF writers so that the always provide the
namespaces explicitly.

Change-Id: I144bab6daee284b424805c1c56d53d385f80b8ee
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>