core: Use ping verb to check connectivity with host
The previous check connectivity used the getCapabilities verb to verify
ovirt-engine is capable of connecting to the host. However, it was used
only to assure the host is responsive and accessible, regardless the
returned data from the getCapabilities.

The patch replaces the use of 'getCapabilities' with 'ping' verb, which
is lighter and should response quicker, with less processing done on the
host.

Change-Id: Ic528aad6a3489d133da4b8c56c07cb6363e0de2f
Signed-off-by: Moti Asayag <masayag@redhat.com>