core: Fix bridgless network recognition
Bridgeless network was not recognized as part of GetVdsCapabilities
call since it has no ports, but instead returns the name of the
interface it is based on.

Fixed the return value of the call to update the interface if it is
provided in the details (this should also work for bridged networks,
since the interface name can be reported on them too).

Change-Id: Iaf8b2b2d9e9d5f16e6b14ecd0f6d01ced4e54ca7
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>