webadmin: LOCALFS storage path regex validation.
LOCALFS storage path is not validated with regex in frontend, being sent
to the backend/vdsm as is, even invalid. On the other hand, adding a NFS
storage, the path is validated in frontend, avoiding extra processing
and quickly/clearly informing user about what's wrong.
This patch adds the LOCALFS path validation in frontend, in order to
standardize according with the NFS regex path validation.

Change-Id: I89a768957974b67754f5a3b78ec8f773e25adc7b
Signed-off-by: Amador Pahim <apahim@redhat.com>