core: UnPowerMock GetoVirtISOsTest
Removed PowerMock dependency from GetoVirtISOsTest and replaced it with
spying the query and using the new MockConfigRule.

In order to achieve this improvment, a slight refactoring to
GetoVirtISOsQuery was required so it will address the Config with lazy
getters instead of storing the config values upon class loading.
A test was added to assure that changing the functionality to lazy
getters does not break the class' behavior.

Change-Id: I3935138a255a5e91750543b35105e88d0e517084
Signed-off-by: Allon Mureinik <amureini@redhat.com>