core: UnPowerMock RemoveAdGroupCommandTest
Removed PowerMock dependency from RemoveAdGroupCommandTest, and replaced
it with spying the command's relevant method.

In order to achieve this improvement a small refactoring was required in
RemoveAdGroupCommand.
CanRemoveAdGroup was turned into a private instance method (instead of a
static public one), and isLastSuperUserGroup was extracted so the
MultiLevelAdministrationHandler dependecy could be mocked away.

Change-Id: Icf5d7eb0e15c92a43b88f7705754ca8238a48d97
Signed-off-by: Allon Mureinik <amureini@redhat.com>