core: UnPowerMock GetUserVmsByUserIdAndGroupsQueryTest
Removed PowerMock depedency from GetUserVmsByUserIdAndGroupsQueryTest,
and replaced it by spying the query's protected methods.

In order to achieve this improvement, some refactoring was required in
GetUserVmsByUserIdAndGroupsQuery and in
GetUserVmsByUserIdAndGroupsQueryTest. First, some protected metods were
extracted to prevent executeQuery() to call VmHandler and ImagesHandler
directly. Second, the test itslef was slightly simplified so it would
actually test the query's flow and not the underlying handlers'
implementations - which is not appropriate in a UNIT test.

Change-Id: I82f99e6deeb0df9b44bfdbfc41f5da675af858e8
Signed-off-by: Allon Mureinik <amureini@redhat.com>